+++ great_expectations/cli/validation_operator.py
@@ -6,6 +6,7 @@ import click
 
 from great_expectations import DataContext
 from great_expectations import exceptions as ge_exceptions
+from great_expectations.cli.mark import Mark as mark
 
 
 from great_expectations.cli.datasource import (
@@ -28,6 +29,7 @@ except ImportError:
     # just fall through
     SQLAlchemyError = ge_exceptions.ProfilerError
 
+
 @click.group()
 def validation_operator():
     """Validation Operator operations"""
@@ -41,6 +43,9 @@ def validation_operator():
     default=None,
     help="The project's great_expectations directory."
 )
+@mark.cli_as_deprecation("""<yellow>In the next major release this command will be deprecated.
+Please consider using `checkpoint list` if you wish to configure a new checkpoint interactively
+  - `checkpoint run` if you wish to run a saved checkpoint</yellow>""")
 def validation_operator_list(directory):
     """List known Validation Operators."""
     try:
@@ -108,6 +113,12 @@ def validation_operator_list(directory):
     "-d",
     default=None,
     help="The project's great_expectations directory.",
+)
+@mark.cli_as_deprecation(
+    """<yellow>In the next major release this command will be deprecated.
+Please consider using either:
+  - `checkpoint new` if you wish to configure a new checkpoint interactively
+  - `checkpoint run` if you wish to run a saved checkpoint</yellow>"""
 )
 def validation_operator_run(name, run_id, validation_config_file, suite, directory):
     # Note though the long lines here aren't pythonic, they look best if Click does the line wraps.